+++ packages/apps/dev-wallet/src/pages/transaction/components/TxTile.tsx
@@ -1,15 +1,8 @@
-import {
-  ITransaction,
-  transactionRepository,
-  TransactionStatus,
-} from '@/modules/transaction/transaction.repository';
+import { ITransaction } from '@/modules/transaction/transaction.repository';
 
 import { shorten } from '@/utils/helpers';
 import {
   MonoBrightness1,
-  MonoCheck,
-  MonoClose,
-  MonoLoading,
   MonoOpenInFull,
   MonoSignature,
   MonoViewInAr,
@@ -20,31 +13,9 @@ import classNames from 'classnames';
 
 import { IPactCommand } from '@kadena/client';
 
-import { useAsync } from '@/utils/useAsync';
 import { Value } from './helpers';
-import {
-  codeClass,
-  failureClass,
-  pendingClass,
-  successClass,
-  txTileClass,
-  txTileContentClass,
-} from './style.css';
-
-export const steps: TransactionStatus[] = [
-  'initiated',
-  'signed',
-  'preflight',
-  'submitted',
-  'failure',
-  'success',
-  'persisted',
-];
-
-export const statusPassed = (
-  txStatus: ITransaction['status'],
-  status: ITransaction['status'],
-) => steps.indexOf(txStatus) >= steps.indexOf(status);
+import { codeClass, txTileClass, txTileContentClass } from './style.css';
+import { getStatusClass, TxPipeLine } from './TxPipeLine';
 
 export const TxTile = ({
   tx,
@@ -60,22 +31,6 @@ export const TxTile = ({
   sendDisabled?: boolean;
 }) => {
   const command: IPactCommand = JSON.parse(tx.cmd);
-  const [contTx] = useAsync(
-    (transaction) =>
-      transaction.continuation?.continuationTxId
-        ? transactionRepository.getTransaction(
-            transaction.continuation?.continuationTxId,
-          )
-        : Promise.resolve(null),
-    [tx],
-  );
-  console.log('tx', tx);
-  const getStatusClass = (status: ITransaction['status']) => {
-    if (statusPassed(status, 'success')) return successClass;
-    if (status === 'failure') return failureClass;
-    if (status === 'initiated') return '';
-    return pendingClass;
-  };
   return (
     <Stack
       flexDirection={'column'}
@@ -91,156 +46,7 @@ export const TxTile = ({
           </Text>
           <MonoBrightness1 className={classNames(getStatusClass(tx.status))} />
         </Stack>
-        {!contTx && statusPassed(tx.status, 'signed') && (
-          <Stack>
-            <Text size="smallest" className={successClass}>
-              <Stack alignItems={'center'} gap={'xs'}>
-                <MonoCheck />
-                Signed
-              </Stack>
-            </Text>
-          </Stack>
-        )}
-        {!contTx && statusPassed(tx.status, 'preflight') && (
-          <Stack>
-            <Text size="smallest" className={successClass}>
-              <Stack alignItems={'center'} gap={'xs'}>
-                <MonoCheck />
-                preflight
-              </Stack>
-            </Text>
-          </Stack>
-        )}
-        {!contTx && statusPassed(tx.status, 'submitted') && (
-          <Stack>
-            <Text size="smallest" className={successClass}>
-              <Stack alignItems={'center'} gap={'xs'}>
-                <MonoCheck />
-                Send
-              </Stack>
-            </Text>
-          </Stack>
-        )}
-        {!contTx &&
-          statusPassed(tx.status, 'submitted') &&
-          (!('result' in tx) || !tx.result) && (
-            <Stack>
-              <Text size="smallest" className={pendingClass}>
-                <Stack alignItems={'center'} gap={'xs'}>
-                  <MonoLoading />
-                  Mining...
-                </Stack>
-              </Text>
-            </Stack>
-          )}
-        {statusPassed(tx.status, 'success') && (
-          <Stack>
-            <Text size="smallest" className={successClass}>
-              <Stack alignItems={'center'} gap={'xs'}>
-                <MonoCheck />
-                Mined{' '}
-                {tx.continuation?.autoContinue && contTx
-                  ? `in chain ${tx.purpose!.data.source as string}`
-                  : ''}
-              </Stack>
-            </Text>
-          </Stack>
-        )}
-        {tx.status === 'failure' && (
-          <Stack>
-            <Text size="smallest" className={failureClass}>
-              <Stack alignItems={'center'} gap={'xs'}>
-                <MonoClose />
-                Failed
-              </Stack>
-            </Text>
-          </Stack>
-        )}
-        {statusPassed(tx.status, 'success') && (
-          <>
-            {tx.continuation?.autoContinue && !contTx && (
-              <Stack>
-                <Text size="smallest" className={pendingClass}>
-                  <Stack alignItems={'center'} gap={'xs'}>
-                    <MonoLoading />
-                    Fetching proof
-                  </Stack>
-                </Text>
-              </Stack>
-            )}
-            {tx.continuation?.autoContinue && tx.continuation.proof && (
-              <Stack>
-                <Text size="smallest" className={successClass}>
-                  <Stack alignItems={'center'} gap={'xs'}>
-                    <MonoCheck />
-                    proof fetched
-                  </Stack>
-                </Text>
-              </Stack>
-            )}
-            {contTx && (
-              <>
-                <Stack justifyContent={'space-between'}>
-                  <Text>cont: {shorten(contTx.hash, 6)}</Text>
-                  <MonoBrightness1
-                    className={classNames(getStatusClass(contTx.status))}
-                  />
-                </Stack>
-                {statusPassed(contTx.status, 'preflight') && (
-                  <Stack>
-                    <Text size="smallest" className={successClass}>
-                      <Stack alignItems={'center'} gap={'xs'}>
-                        <MonoCheck />
-                        preflight
-                      </Stack>
-                    </Text>
-                  </Stack>
-                )}
-                {statusPassed(contTx.status, 'submitted') && (
-                  <Stack>
-                    <Text size="smallest" className={successClass}>
-                      <Stack alignItems={'center'} gap={'xs'}>
-                        <MonoCheck />
-                        Send
-                      </Stack>
-                    </Text>
-                  </Stack>
-                )}
-                {statusPassed(contTx.status, 'submitted') &&
-                  (!('result' in contTx) || !contTx.result) && (
-                    <Stack>
-                      <Text size="smallest" className={pendingClass}>
-                        <Stack alignItems={'center'} gap={'xs'}>
-                          <MonoLoading />
-                          Mining...
-                        </Stack>
-                      </Text>
-                    </Stack>
-                  )}
-                {statusPassed(contTx.status, 'success') && (
-                  <Stack>
-                    <Text size="smallest" className={successClass}>
-                      <Stack alignItems={'center'} gap={'xs'}>
-                        <MonoCheck />
-                        Mined
-                      </Stack>
-                    </Text>
-                  </Stack>
-                )}
-                {contTx.status === 'failure' && (
-                  <Stack>
-                    <Text size="smallest" className={successClass}>
-                      <Stack alignItems={'center'} gap={'xs'}>
-                        <MonoCheck />
-                        Failed
-                      </Stack>
-                    </Text>
-                  </Stack>
-                )}
-              </>
-            )}
-          </>
-        )}
+        <TxPipeLine tx={tx} variant="tile" />
         {tx.status === 'initiated' && (
           <>
             {'exec' in command.payload && (
